+++ evalml/pipelines/components/transformers/samplers/oversamplers.py
@@ -1,11 +1,9 @@
 from evalml.pipelines.components.transformers.samplers.base_sampler import (
-    BaseSampler
+    BaseOverSampler
 )
-from evalml.utils import import_or_raise
-from evalml.utils.woodwork_utils import infer_feature_types
 
 
-class SMOTESampler(BaseSampler):
+class SMOTESampler(BaseOverSampler):
     """SMOTE Oversampler component. Works on numerical datasets only. This component is only run during training and not during predict."""
     name = "SMOTE Oversampler"
     hyperparameter_ranges = {}
@@ -21,47 +19,18 @@ class SMOTESampler(BaseSampler):
             k_neighbors (int): The number of nearest neighbors to used to construct synthetic samples. Defaults to 5.
             n_jobs (int): The number of CPU cores to use. Defaults to -1.
         """
-        error_msg = "imbalanced-learn is not installed. Please install using 'pip install imbalanced-learn'"
-        self.im = import_or_raise("imblearn.over_sampling", error_msg=error_msg)
         parameters = {"sampling_ratio": sampling_ratio,
                       "sampling_ratio_dict": sampling_ratio_dict,
                       "k_neighbors": k_neighbors,
                       "n_jobs": n_jobs}
         parameters.update(kwargs)
-        super().__init__(parameters=parameters,
+        super().__init__("SMOTE",
+                         parameters=parameters,
                          component_obj=None,
                          random_seed=random_seed)
 
-    def fit(self, X, y):
-        """Fits the sampler to the data.
 
-        Arguments:
-            X (ww.DataFrame): Training features
-            y (ww.DataColumn): Target features
-
-        Returns:
-            self
-        """
-        super().fit(X, y)
-        super()._initialize_oversampler(X, y, self.im.SMOTE)
-
-    def fit_transform(self, X, y):
-        """Fit and transform the data using the undersampler. Used during training of the pipeline
-
-        Arguments:
-            X (ww.DataFrame): Training features
-            y (ww.DataColumn): Target features
-
-         Returns:
-            ww.DataTable, ww.DataColumn: Undersampled X and y data
-        """
-        self.fit(X, y)
-        _, _, X_pd, y_pd = self._prepare_data(X, y)
-        X_new, y_new = self._component_obj.fit_resample(X_pd, y_pd)
-        return infer_feature_types(X_new), infer_feature_types(y_new)
-
-
-class SMOTENCSampler(BaseSampler):
+class SMOTENCSampler(BaseOverSampler):
     """SMOTENC Oversampler component. Uses SMOTENC to generate synthetic samples. Works on a mix of nomerical and categorical columns.
        This component is only run during training and not during predict."""
     name = "SMOTENC Oversampler"
@@ -81,41 +50,19 @@ class SMOTENCSampler(BaseSampler):
             k_neighbors (int): The number of nearest neighbors to used to construct synthetic samples. Defaults to 5.
             n_jobs (int): The number of CPU cores to use. Defaults to -1.
         """
-        error_msg = "imbalanced-learn is not installed. Please install using 'pip install imbalanced-learn'"
-        self.im = import_or_raise("imblearn.over_sampling", error_msg=error_msg)
         parameters = {"categorical_features": categorical_features,
                       "sampling_ratio": sampling_ratio,
                       "sampling_ratio_dict": sampling_ratio_dict,
                       "k_neighbors": k_neighbors,
                       "n_jobs": n_jobs}
         parameters.update(kwargs)
-        super().__init__(parameters=parameters,
+        super().__init__("SMOTENC",
+                         parameters=parameters,
                          component_obj=None,
                          random_seed=random_seed)
 
-    def fit(self, X, y):
-        """Fits the sampler to the data.
 
-        Arguments:
-            X (ww.DataFrame): Training features
-            y (ww.DataColumn): Target features
-
-        Returns:
-            self
-        """
-        super().fit(X, y)
-        super()._initialize_oversampler(X, y, self.im.SMOTENC)
-
-    def fit_transform(self, X, y):
-        """Resample the data
-        """
-        self.fit(X, y)
-        _, _, X_pd, y_pd = self._prepare_data(X, y)
-        X_new, y_new = self._component_obj.fit_resample(X_pd, y_pd)
-        return infer_feature_types(X_new), infer_feature_types(y_new)
-
-
-class SMOTENSampler(BaseSampler):
+class SMOTENSampler(BaseOverSampler):
     """SMOTEN Oversampler component. Uses SMOTEN to generate synthetic samples. Works for purely categorical datasets.
        This component is only run during training and not during predict."""
     name = "SMOTEN Oversampler"
@@ -132,34 +79,12 @@ class SMOTENSampler(BaseSampler):
             k_neighbors (int): The number of nearest neighbors to used to construct synthetic samples. Defaults to 5.
             n_jobs (int): The number of CPU cores to use. Defaults to -1.
         """
-        error_msg = "imbalanced-learn is not installed. Please install using 'pip install imbalanced-learn'"
-        self.im = import_or_raise("imblearn.over_sampling", error_msg=error_msg)
         parameters = {"sampling_ratio": sampling_ratio,
                       "sampling_ratio_dict": sampling_ratio_dict,
                       "k_neighbors": k_neighbors,
                       "n_jobs": n_jobs}
         parameters.update(kwargs)
-        super().__init__(parameters=parameters,
+        super().__init__("SMOTEN",
+                         parameters=parameters,
                          component_obj=None,
                          random_seed=random_seed)
-
-    def fit(self, X, y):
-        """Fits the sampler to the data.
-
-        Arguments:
-            X (ww.DataFrame): Training features
-            y (ww.DataColumn): Target features
-
-        Returns:
-            self
-        """
-        super().fit(X, y)
-        super()._initialize_oversampler(X, y, self.im.SMOTEN)
-
-    def fit_transform(self, X, y):
-        """Resample the data
-        """
-        self.fit(X, y)
-        _, _, X_pd, y_pd = self._prepare_data(X, y)
-        X_new, y_new = self._component_obj.fit_resample(X_pd, y_pd)
-        return infer_feature_types(X_new), infer_feature_types(y_new)
